requirements.txt
dj-database-url==0.5.0
Django==2.1.11
django-cors-headers==3.2.0
django-import-export==1.2.0
django-rest-framework-mongoengine==3.4.0
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0
gunicorn==20.0.0
mongoengine==0.18.2
urllib3
requests
nanoid
django-redis